Main-interface
I added the main interface to the game, which includes buttons, tables,
css, and layout. I haven’t completed a lot of the front end jQuery
stuff, but I have messed with a bit of the functions. I’m not sure how
you guys want to handle the backend stuff and how it communicates with
the front end. We can meet and talk about it tomorrow.